LPS-87616 Move code to base class
